requirements.txt
Click==7.0
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
pkg-resources==0.0.0
RPi.GPIO==0.7.0
six==1.13.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
